% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fdidata}
\alias{fdidata}
\alias{fdieff}
\alias{fdilan}
\title{Fishing effort}
\format{dataframe}
\usage{
fdieff

fdilan
}
\description{
Fishing effort in hour and landings in tons by quarter, gear, area, ICES rectangle (and species for landings) for 2014, 2015 and 2016. 
Data were extracted from the JRC data dissemination facility the 17/11/2017.
See \url{https://stecf.jrc.ec.europa.eu/dd/effort/graphs-quarter}
}
\author{
Laurent Dubroca
}
\keyword{datasets}
