% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbathy.R
\name{getbathy}
\alias{getbathy}
\title{This function downloads bathymetry data from EMODNET using wcs service}
\usage{
getbathy(name = "emodnet:mean", xmin = 15, xmax = 20.5, ymin = 30,
  ymax = 32.5)
}
\arguments{
\item{name}{A character vector of the shortname of the variable: mean, mean_atlas_land,mean_multicolour, mean_rainbowcolour}

\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}
}
\value{
A rasterlayer object
}
\description{
This function downloads bathymetry data from EMODNET using wcs service
}
\examples{
  \dontrun{
img<-getbathy("emodnet:mean",-1,1,49,50)
plot(img)
  } 

}
\keyword{Bathymetry}
\keyword{EMODnet}
\keyword{wcs}
