% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethumactpoint.R
\name{gethumactpointall}
\alias{gethumactpointall}
\title{This function downloads all the human activities point data from EMODNET using wfs service}
\source{
\url{http://www.emodnet-humanactivities.eu}
}
\usage{
gethumactpointall(xmin = -1.3, xmax = 0.3, ymin = 49.2, ymax = 49.9)
}
\arguments{
\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}
}
\value{
A data frame
}
\description{
This function downloads all the human activities point data from EMODNET using wfs service
}
\examples{
\dontrun{	
library(REMODnet)
library(mapdata)
library(ggplot2)
ha<-gethumactpointall(-1,1,49,51)
ggplot(ha,aes(x=x,y=y,color=type))+
borders("worldHires",xlim=range(ha$x),ylim=range(ha$y),fill="light grey",colour="light grey")+
geom_point()+
coord_quickmap(xlim=range(ha$x),ylim=range(ha$y))
}
}
\keyword{EMODnet}
\keyword{activities}
\keyword{human}
\keyword{wfs}
