% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethumactpoly.R
\name{gethumactpoly}
\alias{gethumactpoly}
\title{This function downloads one human activities polygon data from EMODNET using wfs service}
\source{
\url{http://www.emodnet-humanactivities.eu}
}
\usage{
gethumactpoly(name = "natura2000", xmin = -1, xmax = 1, ymin = 49,
  ymax = 50)
}
\arguments{
\item{name}{A character vector of the shortname of the variable}

\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}
}
\value{
A spatial polygon data frame
}
\description{
This function downloads one human activities polygon data from EMODNET using wfs service
}
\keyword{EMODnet}
\keyword{activities}
\keyword{human}
\keyword{wfs}
