% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseabed.R
\name{getseabed}
\alias{getseabed}
\title{This function downloads seabed habitat data from EMODNET using wms service}
\source{
https://marinemaps.jrc.ec.europa.eu/
}
\usage{
getseabed(xmin = 15, xmax = 20.5, ymin = 30, ymax = 32.5)
}
\arguments{
\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}

\item{name}{A character vector of the shortname of the variable: mean, mean_atlas_land,mean_multicolour, mean_rainbowcolour}
}
\value{
A rasterlayer object
}
\description{
This function downloads seabed habitat data from EMODNET using wms service
}
\examples{
  \dontrun{
img<-getseabed(-1,1,49,50)
rasterVis::levelplot(img,par.settings=rasterVis::rasterTheme(region=img@data@attributes[[1]]$color))
  } 

}
\keyword{EMODnet}
\keyword{Seabed}
\keyword{habitat}
\keyword{wms}
