% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMODnet.R
\docType{package}
\name{REMODnet}
\alias{REMODnet}
\alias{REMODnet-package}
\title{REMODnet}
\description{
A package to gather and analyses data from EMODnet
}
\details{
The European Marine Observation and Data Network (EMODnet) consists of more than 150 organisations assembling marine data, products and metadata to make these fragmented resources more available to public and private users relying on quality assured, standardised and harmonised marine data which are interoperable and free of restrictions on use. EMODnet is currently in its second development phase with the target to be fully deployed by 2020.

This package...
}
