% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data.R, R/seabedcolorbar.R
\docType{data}
\name{seabedcolorbar}
\alias{seabedcolorbar}
\alias{seabedcolorbar}
\title{Seabed colorbar}
\format{dataframe

Columns:

RGB = RGB code
code = seabed code
Description = description of seabed classes
RGB2 = internet RGB code
color = color in R format}
\source{
https://marinemaps.jrc.ec.europa.eu/

https://marinemaps.jrc.ec.europa.eu/
}
\usage{
seabedcolorbar

seabedcolorbar
}
\description{
Seabed colorbar

Seabed colorbar
}
\keyword{datasets}
