% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethumactpoint.R
\name{gethumactpoint}
\alias{gethumactpoint}
\title{This function downloads one human activities point data from EMODNET using wfs service}
\source{
\url{http://www.emodnet-humanactivities.eu}
}
\usage{
gethumactpoint(name = "shellfish", xmin = -1000, xmax = 1000,
  ymin = -1000, ymax = 1000)
}
\arguments{
\item{name}{A character vector of the shortname of the variable}

\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}
}
\value{
A data frame
}
\description{
This function downloads one human activities point data from EMODNET using wfs service
}
\keyword{EMODnet}
\keyword{activities}
\keyword{human}
\keyword{wfs}
