% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseabed.R
\name{getseabed}
\alias{getseabed}
\title{This function downloads seabed habitat data from EMODNET using wms service}
\source{
\url{https://marinemaps.jrc.ec.europa.eu/}
}
\usage{
getseabed(xmin = -1.3, xmax = 0.3, ymin = 49.2, ymax = 49.9)
}
\arguments{
\item{xmin}{A numeric vector of the lower longitude of the spatial area}

\item{xmax}{A numeric vector of the max longitude of the spatial area}

\item{ymin}{A numeric vector of the lower latitude of the spatial area}

\item{ymax}{A numeric vector of the max latitude of the spatial area}
}
\value{
A rasterlayer object
}
\description{
This function downloads seabed habitat data from EMODNET using wms service
}
\examples{
  \dontrun{
img<-getseabed(-1.3,0.3,49.2,49.9)
rasterVis::levelplot(img,par.settings=rasterVis::rasterTheme(region=img@data@attributes[[1]]$color))
  } 

}
\keyword{EMODnet}
\keyword{Seabed}
\keyword{habitat}
\keyword{wms}
