% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seabedcolorbar}
\alias{seabedcolorbar}
\title{Seabed colorbar}
\format{dataframe}
\usage{
seabedcolorbar
}
\description{
A colorbar used to tranforsm seabed map in seabed data.
See \url{https://marinemaps.jrc.ec.europa.eu}.
Columns:
RGB (RGB code), code (seabed code),
Description (description of seabed classes),
RGB2 (internet RGB code), color (color in R format).
}
\author{
Laurent Dubroca
}
\keyword{datasets}
